refactor: move NamecoinResolveState to commons
Move NamecoinResolveState sealed class to commons.service.namecoin
for KMP shared access. Uses quartz NamecoinNostrResult which is
already multiplatform.
